requirements.txt
asgiref==3.7.2
class-registry==2.1.2
Django==5.0.3
django-filter==23.5
djangorestframework==3.14.0
filters==1.3.2
pillow==10.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
